added external fetch for falling books
Now similar to the question mark quotes the page data list is generated via an external file.

Added a semi random generator for selecting page_data indexes

Patched a bug with some times the search for an maching element in falling_books comes back -1. I added a check for it but it should not be -1 one in the first place, which is strange behavior.